LayoutTests/gamepad/gamepad-vibrationActuator-type-expected.txt

Tests for vibrationActuator.type

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gamepad.vibrationActuator.type is "dual-rumble"
PASS successfullyParsed is true

TEST COMPLETE
